Fix issues pointed by Ugo
